Moved GPU implementation of Dahlquist problem to its own file to avoid importing cupy when accessing CPU version
